+++ terraform/container-registry.tf
@@ -20,6 +20,22 @@ resource "azurerm_container_registry" "j1dev" {
   sku                 = "Basic"
 }
 
+resource "azurerm_monitor_diagnostic_setting" "j1dev_cont_reg_diag_set" {
+  name               = "j1dev_cont_reg_diag_set"
+  target_resource_id = azurerm_container_registry.j1dev[0].id
+  storage_account_id = azurerm_storage_account.j1dev.id
+
+  log {
+    category = "ContainerRegistryLoginEvents"
+    enabled  = true
+
+    retention_policy {
+      enabled = true
+      days    = 1
+    }
+  }
+}
+
 resource "azurerm_container_registry_webhook" "j1dev" {
   count               = local.container_registry_resource_count
   name                = "j1dev"